+++ src/demo_evm_uniswap/models/position.py
@@ -9,7 +9,7 @@ from demo_evm_uniswap.models.abi import get_abi
 async def position_validate(
     ctx: HandlerContext, contract_address: str, position_id: int, position: models.Position,
 ) -> None:
-    web3 = ctx.get_evm_node_datasource('subsquid').web3
+    web3 = ctx.get_evm_node_datasource('evm_node').web3
     manager = web3.eth.contract(
         address=to_checksum_address(contract_address),
         abi=get_abi('position_manager.abi'),